requirements.txt
blinker==1.9.0
bottle==0.13.3
cffi==1.17.1
click==8.1.8
clr_loader==0.2.7.post0
colorama==0.4.6
Flask==3.1.1
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
proxy_tools==0.1.0
pycparser==2.22
pythonnet==3.0.5
pywebview==5.4
typing_extensions==4.13.2
Werkzeug==3.1.3
zipp==3.21.0
